requirements.txt
asgiref
autopep8
django==2.2.10
pycodestyle
pytz
sqlparse
Unipath
dj-database-url
python-decouple
gunicorn
whitenoise
django-debug-toolbar
